Remove dependency on old conduit
The dependencies on bzlib-conduit and lzma-conduit pinned this
package to an old version of conduit that was getting more difficult
to install.  This change gets rid of conduit and just uses the
process package to invoke decompression filters as separate
processes.

The package now depends on gunzip, bunzip2, and unxz.